+++ core/core-impl/src/main/java/org/openbaton/nfvo/core/api/VirtualNetworkFunctionManagement.java
@@ -6,6 +6,8 @@ import org.openbaton.nfvo.repositories.VnfPackageRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Service;
 
@@ -14,6 +16,7 @@ import org.springframework.stereotype.Service;
  */
 @Service
 @Scope
+@ConfigurationProperties
 public class VirtualNetworkFunctionManagement implements org.openbaton.nfvo.core.interfaces.VirtualNetworkFunctionManagement {
 
     @Autowired
@@ -21,6 +24,16 @@ public class VirtualNetworkFunctionManagement implements org.openbaton.nfvo.core
     @Autowired
     private VnfPackageRepository vnfPackageRepository;
     private Logger log = LoggerFactory.getLogger(this.getClass());
+    @Value("${vnfd.vnfp.cascade.delete:false}")
+    private boolean cascadeDelete;
+
+    public boolean isCascadeDelete() {
+        return cascadeDelete;
+    }
+
+    public void setCascadeDelete(boolean cascadeDelete) {
+        this.cascadeDelete = cascadeDelete;
+    }
 
     @Override
     public VirtualNetworkFunctionDescriptor add(VirtualNetworkFunctionDescriptor virtualNetworkFunctionDescriptor) {
@@ -33,8 +46,10 @@ public class VirtualNetworkFunctionManagement implements org.openbaton.nfvo.core
         VirtualNetworkFunctionDescriptor virtualNetworkFunctionDescriptor = vnfdRepository.findFirstById(id);
         log.info("Removing VNFD: " + virtualNetworkFunctionDescriptor.getName());
         vnfdRepository.delete(virtualNetworkFunctionDescriptor);
-        log.info("Removing vnfPackage with id: " + virtualNetworkFunctionDescriptor.getVnfPackageLocation());
-        vnfPackageRepository.delete(virtualNetworkFunctionDescriptor.getVnfPackageLocation());
+        if (cascadeDelete) {
+            log.info("Removing vnfPackage with id: " + virtualNetworkFunctionDescriptor.getVnfPackageLocation());
+            vnfPackageRepository.delete(virtualNetworkFunctionDescriptor.getVnfPackageLocation());
+        }
     }
 
     @Override
